Refactor: fix intellij inspection issues and move commands

+++ src/main/java/com/alistats/discorki/util/ColorUtil.java
@@ -18,7 +18,7 @@ public class ColorUtil {
     public static int generateRandomColorFromString(String string) {
         int sum = 0;
         for (int i = 0; i < string.length(); i++) {
-            sum += (int) string.charAt(i);
+            sum += string.charAt(i);
         }
         Random random = new Random();
         random.setSeed(sum);
